+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fixed bug where very long WDLs could not be saved in Db and would fail.
 - Cancelled runs will now be downloaded and available via the "get" command (useful for debugging).
 - Include the stdout.submit in the errors report (previously only included stderr.submit file)
+- Include transfer failure error message in run log
 
 ## 2.8.5 (2022-5-24) Summary
 ### User Facing Changes